Add a password policy for new passwords, including tests.
The password policy is implemented in PasswordService using the Passay library.
The policy affects passwords created on createAccount.gsp and passwordReset.gsp.
Added i18n text to explain the password policy on the two view pages.
Existing passwords are not affected.
Standardised password length to 8 characters.
Includes some changes to the password generation and validation process to account for the password policy.